Extracting tests into individual files by subject
